replace isMac function
